third_party/blink/web_tests/platform/generic/external/wpt/websockets/Create-on-worker-shutdown.any-expected.txt

This is a testharness.js-based test.
FAIL WebSocket created after a worker self.close() assert_equals: Closed WebSocket created on worker shutdown is in closing state. expected 2 but got 3
Harness: the test ran to completion.
